Changed int4 typedefs to use POSIX uint32_t
git-svn-id: http://svn.osgeo.org/postgis/trunk@2302 b70326c6-7e19-0410-871a-916f4a2858ee

+++ lwgeom/lwgparse.c
@@ -7,6 +7,7 @@
  */
 #include <string.h>
 #include <stdio.h>
+#include <stdint.h>
 
 #include "liblwgeom.h"
 #include "wktparse.h"
@@ -22,7 +23,7 @@
 void set_zm(char z, char m);
 void close_parser(void);
 
-typedef unsigned long int4;
+typedef uint32_t int4;
 
 typedef struct tag_tuple tuple;
