+++ lwgeom/wktunparse.c
@@ -14,6 +14,7 @@
 /* TO get byte order */
 #include <sys/types.h>
 #include <sys/param.h>
+#include <stdint.h>
 
 #include "liblwgeom.h"
 #include "wktparse.h"
@@ -21,7 +22,7 @@
 
 /*-- Typedefs ---------------------------------------------- */
 
-typedef unsigned long int4;
+typedef uint32_t int4;
 typedef uchar* (*outfunc)(uchar*,int);
 typedef uchar* (*outwkbfunc)(uchar*);
 
@@ -610,6 +611,9 @@ unparse_WKB(uchar* serialized, allocator alloc, freeor free, char endian, size_t
 
 /******************************************************************
  * $Log$
+ * Revision 1.21  2006/02/03 09:52:14  strk
+ * Changed int4 typedefs to use POSIX uint32_t
+ *
  * Revision 1.20  2006/01/09 15:12:02  strk
  * ISO C90 comments
  *
